requirements.txt
yara-python>=4.5
psutil>=7.0